Refactor tts code into its own router
